Update outbound for sell

+++ src/main/java/com/example/hrm_be/models/dtos/OutboundDetail.java
@@ -30,7 +30,7 @@ public class OutboundDetail {
 
   Product product;
 
-  Integer quantity;
+  BigDecimal quantity;
 
   UnitOfMeasurement unitOfMeasurement;
 }
