+++ src/main/java/com/example/hrm_be/models/dtos/OutboundProductDetail.java
@@ -34,7 +34,7 @@ public class OutboundProductDetail {
 
   List<Batch> batches;
 
-  Integer outboundQuantity;
+  BigDecimal outboundQuantity;
 
   BigDecimal price;
