+++ src/main/java/com/example/hrm_be/models/entities/OutboundDetailEntity.java
@@ -48,7 +48,7 @@ public class OutboundDetailEntity extends CommonEntity {
   UnitOfMeasurementEntity unitOfMeasurement;
 
   @Column(name = "quantity")
-  Integer quantity;
+  BigDecimal quantity;
 
   @Column(name = "price")
   BigDecimal price;
